Added authentication using jwtTokens

+++ backend/controller/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -15,7 +16,7 @@ const setGoals = asyncHandler(async (req, res) => {
     throw new Error("please enter a text");
   }
 
-  const goal = await Goal.create({ text: req.body.text });
+  const goal = await Goal.create({ user: req.user.id, text: req.body.text });
   res.status(200).json(goal);
 });
 
@@ -24,6 +25,19 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged in user matches the goal user.
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -39,6 +53,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged in user matches the goal user.
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user not authorized");
+  }
   await Goal.deleteOne({ _id: req.params.id });
   res.status(200).json({ id: req.params.id });
 });
